requirements.txt
## Code quality
pytest==7.0.1
black==22.3.0

## Standard science stack
h5py
matplotlib
numpy==1.23
scipy

## Standard utils
tqdm

## Deep learning
einops
git+https://github.com/dgasmith/opt_einsum.git@1a984b7b75f3e532e7129f6aa13f7ddc3da66e10 # Un-released latest master (pinned for repro); the latest released version 3.3.0 has bugs in shape-based contraction path computations (used for einsum)
torch==2.0.1
torch_geometric==2.4.0

## Experiment management
hydra-core
mlflow==2.4.1 # avoid database version issues
optuna

## Project-specific: general
clifford==1.4.0
xformers==0.0.20
